move localization helper
