+++ plugin-hrm-form/src/translations/1baseLanguage/es.json
@@ -1,3 +1,23 @@
 {
-  "CancelButton": "Cancelar"
+  "TaskInfoPanelContent": "<h1>CONTEXTO DE TAREA</h1><h2>Tipo de tarea</h2><p>{{task.channelType}}</p><h2>Fecha creación de tarea</h2><p>{{task.dateCreated}}</p><h2>Prioridad de tarea</h2><p>{{task.priority}}</p><h2>Cola de tarea</h2><p>{{task.queueName}}</p><h2>Identificador de tarea</h2><p>{{task.taskSid}}</p><h2>Identificador de reserva</h2><p>{{task.sid}}</p><hr /><h1>CONTEXTO DE CLIENTE</h1><h2>Nombre / número de teléfono</h2><p>{{task.defaultFrom}}</p><h2>País</h2><p>{{task.attributes.caller_country}}</p><hr /><h1>ADDONS</h1><p>    No hay add-ons activados.    Para mejorar su experiencia, visite</p><a href=\"https://www.twilio.com/marketplace/add-ons\" target=\"blank\">Twilio Marketplace</a>",
+  "SupervisorTaskInfoPanelContent": "<h1>CONTEXTO DE TAREA</h1><h2>Tipo de tarea</h2><p>{{task.channelType}}</p><h2>Fecha creación de tarea</h2><p>{{task.dateCreated}}</p><h2>Prioridad de tarea</h2><p>{{task.priority}}</p><h2>Cola de tarea</h2><p>{{task.queueName}}</p><h2>Identificador de tarea</h2><p>{{task.taskSid}}</p><h2>Identificador de reserva</h2><p>{{task.sid}}</p><hr /><h1>CONTEXTO DE CLIENTE</h1><h2>Nombre / número de teléfono</h2><p>{{task.defaultFrom}}</p><h2>País</h2><p>{{task.attributes.caller_country}}</p><hr /><h1>ADDONS</h1><p>    No hay add-ons activados.    Para mejorar su experiencia, visite</p><a href=\"https://www.twilio.com/marketplace/add-ons\" target=\"blank\">Twilio Marketplace</a>",
+  "TaskHeaderStatusAccepted": "En vivo | {{helper.durationSinceUpdate}}",
+  "TaskHeaderWebChatStatusAccepted": "{{#if helper.chat.isCustomerOnline}} {{{icon name='GreenIndicator'}}} {{else}} {{{icon name='GreyIndicator'}}} {{/if}} En vivo | {{helper.durationSinceUpdate}}",
+  "TaskHeaderEndChat": "FINALIZAR CHAT",
+  "ChatWelcomeText": "Conversación iniciada",
+  "AgentStatusAvailable": "Disponible",
+  "AgentStatusBusy": "Ocupado",
+  "AgentStatusUnavailable": "No disponible",
+  "AgentStatusOffline": "Offline",
+
+  "TranslateButtonAriaLabel": "Cambiar idioma",
+
+  "CallType-child": "Menor llamando por su cuenta",
+  "CallType-caller": "Alguien llamando por un menor",
+  "CallType-silent": "Silencioso",
+  "CallType-blank": "Blanco",
+  "CallType-joke": "Broma",
+  "CallType-hangup": "Cuelga",
+  "CallType-wrongnumber": "Número Equivocado",
+  "CallType-abusive": "Abusivo"
 }
